refactor(htmlhint): use umd module in tests
