+++ package.json
@@ -24,7 +24,7 @@
   "scripts": {
     "commit": "npx git-cz",
     "prettier": "prettier --write ./**/*.{js,json,md} --ignore-path ./.prettierignore",
-    "test": "grunt exec:test",
+    "test": "mocha",
     "build": "webpack",
     "travis-deploy-once": "travis-deploy-once",
     "semantic-release": "semantic-release"
